# specs/001-phase2-full-stack-web/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Phase 2 Todo Application API
  description: REST API for todo application with user authentication and task management
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server (TBD)

tags:
  - name: Authentication
    description: User signup, signin, and token management
  - name: Tasks
    description: CRUD operations for tasks (requires authentication)

paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Create new user account
      description: Register a new user with email and password
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              email: alice@example.com
              password: securepass123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Email already registered
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign in to existing account
      description: Authenticate user and receive JWT tokens
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
            example:
              email: alice@example.com
              password: securepass123
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Invalid credentials

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks for authenticated user
      description: Returns all tasks owned by the current user
      operationId: getTasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tasks
      summary: Create new task
      description: Create a new task for the authenticated user
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: Buy groceries
              description: Get milk, eggs, bread
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get single task by ID
      description: Returns task details if owned by authenticated user
      operationId: getTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update task title and/or description (ownership validated)
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: Buy groceries and supplies
              description: Get milk, eggs, bread, and cleaning products
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete task (ownership validated)
      operationId: deleteTask
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{task_id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion status
      description: Toggle task between incomplete and complete (ownership validated)
      operationId: toggleTaskCompletion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header

  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      description: UUID of the task
      schema:
        type: string
        format: uuid
      example: 660e8400-e29b-41d4-a716-446655440001

  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: alice@example.com
        password:
          type: string
          minLength: 8
          description: Password (minimum 8 characters)
          example: securepass123

    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: securepass123

    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token received from signin/signup
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15-minute expiration)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token (7-day expiration)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          default: bearer
          example: bearer

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (required, max 500 chars)
          example: Buy groceries
        description:
          type: string
          nullable: true
          description: Optional task description
          example: Get milk, eggs, bread

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          nullable: true
          description: Updated task title
          example: Buy groceries and supplies
        description:
          type: string
          nullable: true
          description: Updated task description
          example: Get milk, eggs, bread, and cleaning products

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task UUID
          example: 660e8400-e29b-41d4-a716-446655440001
        user_id:
          type: string
          format: uuid
          description: Owner's user UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Get milk, eggs, bread
        status:
          type: string
          enum:
            - incomplete
            - complete
          example: incomplete
        created_at:
          type: string
          format: date-time
          example: '2025-12-29T10:30:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-12-29T10:30:00Z'

    Error:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Task not found

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  anyOf:
                    - type: string
                    - type: integer
                description: Location of the error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
          example:
            - loc: ['body', 'title']
              msg: 'field required'
              type: 'value_error.missing'

  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: No token provided
              value:
                detail: Authentication required
            invalid_token:
              summary: Invalid or expired token
              value:
                detail: Token expired